NestedForm: Validate dynamically added form elements.
The basic strategy: On the server, compile a list of NestedForm
"blueprint" elements and their validators. On the client, when we're
asked to validate an element we don't recognize, see if it matches
NestedForm's pattern of elements generated from blueprints, and copy
the blueprint validators if so.

Two complications of the basic strategy:

1. NestedForm generates the blueprints *after* the form, so we can't
easily hook into form_for and add the blueprint validators at that
point. Instead, we wrap nested_form_for, and then check the validators
array for element names that match the blueprint pattern. It would be
nice if we could do this with less knowledge of NestedForm internals.

2. On the client side, we copy the blueprint validators into the
normal validators object. This is because only validators are passed
down into the validation process. Blueprint element names are unlikely
to cause a conflict with normal element names, so this probably isn't
a problem in practice, but it would be better if we could keep the
blueprint validators separate.